allow JenkinsLocationConfiguration adminAddress to be null
The default value for the adminAddress field is null.  The setAdminAddress
setter method prevents this field from being [re]set to null as it does not
validate its string param as being not null before attempting string comparison
operations.  This simple fix skips string comparison checks if the string is
null.