JenkinsTest fixes
runScriptOnOfflineComputer has always been dodgy from what I can see. It was relying on the slave not being connected at the time the doScript call happened. I tested this on a vanilla master and it's easy to make this test fail  -just put a loop in after the slave is created, waiting on the slave channel to be connected before making the doScript call.

This change makes sure the slave has connected and disconnected before continuing.