[FIXED JENKINS-30039] Builds executed concurrently even with single execution selected
The issue here is a 'race' between the Executor to create the Run and the Queue to schedule the next buildable.

Because the Queue schedules multiple buildables at a time, and Exector has to wait for the Queue lock to create the Run,
Job.isLogUpdated() will not detect the scheduled WorkUnit if two build requests for the same Job are both buildable
at the same maintain().

This change allows the job to reflect the fact that another request is being started but has not created the SubTask for
the Run just yet.

There is the potential issue of scalability for beyond 1000's of executors. If we reach that point a better solution would
be to have a Set of scheduled but not created work units maintained in a Concurrent collection by Queue and let Job check that
Set. On the other hand, that is complexity without evidence that it is required. On that basis we go for the simpler loop.
(The Set alternative would also require a tidy-up thread to clear out any entries that disappeared before we recorded them
as having created their SubTask... or perhaps Queue.maintain() could do that tidy-up... but still more complex without evidence