[JENKINS-29547] plugins can't set an assigned node of 'master'
Plugins need to use the setter/getter pair, because the 'assignedNode'
field is final.  But the AbstractProject uses different functions to read
and write the value, and didn't previously follow the same invariants as
the get/set pair.
getAssignedLabelString() for example, treats a true 'canRoam' the same as
a null 'assignedNode' field, but these represent different things.
This patch fixes getAssignedLabelString, so that users viewing the config
page will see the right string.