Update getAssignedLabelString test - 'master' is no longer special case
These test cases were submitted in a large batch in 2013, so it's not
clear that they were testing anything other than the code "as written"
from before.  I am not seeing developer intent here, so this updates the
test to match the new logic that works better on the config page.