[FIXED JENKINS-31235] fix createItem/copy API args
Previous to this change, a "Content-Type" HTTP header was required but
ignored for the createItem?mode=copy REST API call.  This header is
meaningless for the create-from-copy operation as the content is sourced
internally rather than provided by the request (there is no
user-supplied 'content' whose type needs describing in this case).  With
this change, this call will succeed without the Content-Type header.