Remove broken XStream2 test of legacy custom ConcurrentHashMapSerialiation
This one requires some explanation: there is a bug
As custom serialization of ConcurrentHashMaps was removed circa 2010
(see eb51d46d56369915ae4031dc167322ad126f0be6)

We can assume that testing conversion is no longer a high priority.
Root cause appears to be a subtle change in the XStream attribute access behavior when run under JDK8 only.
This appears to be the only case that triggers it.