[Fixed JENKINS-33304] prevent bad characters in usernames.
The container authentication can pass us whatever it feels like as a user
name.  The User class and the IdStrategy assume a username is safe for use
on the filesystem.
Mostly this is the case - but some 3rd party integrations such as
Waffle and AD this is not true as we can be passed users of the form
DOMAIN\user

This causes several things to break and there are reports of chrome and IE
not working correctly (although firefor is reported to be ok) with passing
\.
The other areas that do not work where the url escaping of the \ which
then was being munged to an underscore rather than a backslash causing things
like my views in the upper right hand drop down to not work correctly.

Whilst a much bigger fix could be done the scale of the problem is larger
as it requires symetricall remapping (IdStrategy.filenameOf() <->
IdStrategy.idFromFileName()) and the User to keep track of the original
username and the munged user name, hence this is a little small hack that
does just about enough.

One area that is known to be odd now - is that the users name in the API
Token will show the munged name rather than the correct name, however I
fell that users will likely know their username - at least now they can
see the token which before you had to know how to get to it via the /me
URL.