Fixing a test
Previously, when the 2nd `e.click()` is called, the page has already transitioned to another page, so the anchor objects have become invalid.

HtmlUnit somehow still manages to follow a link, except when there's some JavaScript attached to the onclick() handler it's going to execute in a strange environment. Some management link requires a POST, and that was done via JavaScript. So that broke the test.