Revert "[JENKINS-32190] Make UsageStatistics#isDue check Jenkins init level before generating stats."
This reverts commit d8a13ac07c4357eb74ab45a03955c0db430e3aa9.

I noticed a strange drop in usage statistics from the mid Jan 2016,
which corresponds to the time this change was released in the wild,
and further drop when LTS 1.642.2 is released with this change at
the end of Feb.

I spent some time trying to determine how this might cause instances
to stop sending data, but I couldn't come up with any. I left those
notes in JENKINS-32190. There was some strange behaviours, but none
explains what this is.

So I'd like to back out this change and see if that makes any impact
on the stats collection. If it backs up, then we know empirically
this change is to blame, so that would justify spending further
effort or come up with a fix from a different angle.

(cherry picked from commit b2013edbf7a4d40df2bc05a1e219d8599e642961)