Ensure the "Getting Started" dialogue title is consistent between views
The different views are distinguished by the header (h1) tags, for example
"Customize Jenkins" or "Jenkins is ready!" and I find it jarring to have the
dialogue title change.

The context of the whole series of views in the dialogue is "Getting Started" so
I think the dialogue title should reflect that scoping of actions.