[FIXED JENKINS-33693] Do not hardcode .bat extension.
As the default maven command is on the path we can not hard code the
command to mvn.bat as since maven 3.something it has been mvn.cmd

As this is on the path we can just use mvn and let Windows do its thing
with PATHEXT which will has both .BAT and .CMD by default.

This fixes ATH failures when the ATH is run on windows.

(cherry picked from commit 490adc9d0f4b0ed2c7e3183b18b136c006953824)