[FIXED JENKINS-33848] - Do not try to resolve empty strings in Items:fromNameList()
ReverseBuildTrigger autocompletes lists of jobs by delimiters when you select items, so sometimes you get the "Jib1, Job2, " in the form submission. Then ReverseBuildTrigger invokes Items::fromNameList() in order to resolve the dependency graph.

In this method Jenkins splits the string, trims it and then tries to resolve jobs. In such case an empty string is being passed to the resolution code. It may be unsafe since items (e.g. folders) may somehow override and mishandle the logic.

It would be useful to avoid such resolution attempts.