Fixed a test regression
Previously, wa.contextDestroyed(null) interrupts the Jenkins.<init>
thread, and it was the 2nd reactor call that was throwing InterruptedException
from this call. Now that I removed that 2nd reactor call, the init
thread never actually fails.

The intent of the test case is to make sure the boot failure hook is
invoked, so even with this change it covers the original intent.