Use DirectoryStream paths in java 7 to traverse the users.
With the move to Java 7 the nio2 library is made available for usage. The DirectoryStream api allows for a more effeicient method of walking the directory tree structure. This fixes a potential performance issue where there are several thousand users in the $JENKINS_HOME/users directory which are still on the legacy format for user ids. This uses the NIO2 methods for doing the tree walk and moving the directories into the correct locations.