[JENKINS-18435] Tell browsers not to cache or try to autocomplete forms
Setting autocomplete to "off" will tell browsers not to try to autofill
forms during history navigation. Since this seems to be a prominent
problem in Jenkins I feel it is best to have it default to "off" and let
individual inputs or pages opt-in to enable it instead
to indicate they can handle it.
The standard login page does not seem to use the taglib form tag
so it is not affected, even so, all modern browsers ignores
the autocomplete attribute on username and password fields anyway.