[FIXED JENKINS-37731] Change symbol and constructor for SCMTrigger.
Using "scm" as the @Symbol led to collisions with the "scm" global
variable in Pipeline, so that had to go. Changed it to "pollScm". Also
moved the @DataBoundConstructor to the single-parameter constructor,
deprecated the two-parameter constructor, created a @DataBoundSetter
for ignorePostCommitHooks, and added a getter for "scmpoll_spec" to
fix snippet generator and general syntactic goodness.

So now, with this, it's possible to do:

properties([
  pipelineTriggers([
    pollScm('@daily')
  ])
])

Tada.

Tests for the snippet generator and job property stuff will be landing
in a separate PR for workflow-multibranch.