Update XStream drive to improve performance in xml serilization/deserialization
According XStream FAQ (http://x-stream.github.io/faq.html#Scalability):

XStream is a generalizing library, it inspects and handles your types on
the fly. Therefore it will normally be slower than a piece of optimized
Java code generated out of a schema. However, it is possible to increase
the performance anyway:

* Write custom converters for those of your types that occur very often in
your XML.
* Keep a configured XStream instance for multiple usage. Creation and
initialization is quite expensive compared to the overhead of XStream
when calling marshall or unmarshal.
* Use Xpp3 or StAX parsers.

So, I decided to move from old Xpp to new Xpp3.