Add .gitattributes to embed standard IN the repo
That file will avoid users to file PR with wrong newline encoding.
I've tried to declare what seemed to be the most common extensions. It's
not strictly required, since the first line and the Git integrated
heurisitics would work just fine in general. But we're just adding a bit
more info here to lower chances to screw up.