[SECURITY-343] Restrict API access to NodeMonitor data
Switch to requiring EXTENDED_READ permission for access to NodeMonitor
data. It might theoretically be better if we had more granular
permissions on each NodeMonitor, but that's a bigger change, and since
EXTENDED_READ is already what we use to limit access to serving the
computer's config.xml, I think this is appropriate.