never allow a flyweight task to execute on a restricted node
A previous commit 55bf88329027 ("If every node is restricted to tied
jobs only, Matrix build jobs can never start.", 2013-06-18) made the
decision to sometimes allow flyweight tasks to execute on an exclusive
node. The condition chosen here was that the master executor must either
have <1 executor or be marked exclusive. Unfortunately this means that
sometimes without an obvious explanation, a flyweight task might choose
to pick an exclusive node, even if some other slave node besides the
master was not marked as exclusive.

Revert this fix so that the exclusive nodes remain exclusive. If the
original goal is desirable, an alternative approach would be to check
*every* node and only allow flyweight tasks on an exclusive node if
every node is marked exclusive, rather than only bothering to check
master.

[JENKINS-5076]
[JENKINS-23459]