[JENKINS-5076] only execute flyweight task on master if all nodes are exclusive
Currently, if the master node is marked as exclusive, or has fewer than 1
executor, we will allow flyweight tasks to execute on random exclusive
nodes. This is a good thing, because we want to avoid blocking all
execution of flyweight tasks with a fairly trivial configuration,
especially since most users will not be aware of the flyweight tasks.

However, this solution accidentally allowed the execution of a flyweight
task on exclusive nodes even when there are nodes which are marked as
non-exclusive.

We will fix this with a new solution. In the case when we have a flyweight
task and the master node is marked as exclusive, then we first check
whether every node is exclusive. If we find at least one non-exclusive
node, then we will prefer it instead of the master. However, if we are
unable to find such a node, then we will allow the master to execute this
flyweight task.

This solution neatly solves the original problem, as we will not create a
situation where flyweight matrix or pipeline tasks are not executed.
Additionally, we avoid the issue of executing flyweight tasks on exclusive
nodes unnecessarily. Finally, we now only execute on the master node
rather than on an arbitrary exclusive node, which should match users
general expectations.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>