[JENKINS-5076] avoid running flyweight tasks on exclusive nodes
Since commit 55bf88329027 ("If every node is restricted to tied jobs
only, Matrix build jobs can never start.", 2013-07-03) we've had
a workaround which enables flyweight tasks to run even on nodes marked
as exclusive.

This patch was included because of a possibility that a Jenkins
administrator configured all slaves and the master as exclusive, which
prevented workflow or matrix job flyweight tasks from starting.

Unfortunately, the original solution was too naive. Essentially, if the
master node was marked as exclusive, and had zero executors, then it
would bypass the restriction and allow the flyweight task to run despite
the node being marked as restricted.

Since we cannot simply revert the original change, expand the check so
that we only bypass the restriction if (a) the master is exclusive *and*
(b) every other slave node is also exclusive. Additionally, drop the
check on whether the master node has enough executors, since flyweight
tasks do not consume any executors anyways.

In other words, abide by the exclusivity restrictions for flyweight
tasks unless there is no other option available.

Add a NodeTest class with a couple of tests to verify our assumptions
and prevent future regressions.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>