[FIXED JENKINS-38005] Properly log failure due to empty archive in Pipeline
Since the job result is probably still null at the time that archiving
occurs, we weren't actually logging anything in Pipeline jobs in that
case. That was not ideal. So instead, let's log things, and also let's
update the error message to include a tip to use the allowEmptyArchive
option.